Added fileUpload and fileRetrieval for middleware
