+++ server/routes/api/admin.js
@@ -3,7 +3,8 @@ const File = require('../../models/Files');
 var requireRole = require('../../middleware/Token').requireRole;
 var diskStorage = require('../../middleware/fileStorage').diskStorage;
 var addDirectory = require('../../middleware/fileStorage').addDirectory;
-var path = require('path');
+var fileUpload = require('../../middleware/fileStorage').fileUpload;
+var retrieveFile = require('../../middleware/fileStorage').retrieveFile;
 var dir = process.cwd() + '/../temp';
 var keyName = "inputFile" //Change according to your key name for file
 
@@ -97,108 +98,21 @@ module.exports = (app) => {
 
     var upload = diskStorage(dir); 
 
-    app.post('/api/admin/file', upload.single(keyName), requireRole("admin"), function (req, res) {
+    app.post('/api/admin/upload', upload.single(keyName), requireRole("admin"), fileUpload ,function (req, res) {
         if (!req.file) {
             return res.status(400).send({
                 success: false,
                 message: "Error: File not recieved"
             });
         }
-
-        File.find({
-            user_id: req.user_id,
-            originalname: req.file.originalname
-        }, function (err, files) {
-            if (err) {
-                return res.status(500).send({
-                    success: false,
-                    message: "Error: Server error"
-                });
-            }
-            else if (files.length > 0) {
-                return res.status(400).send({
-                    success: false,
-                    message: "Error: File is already entered by user."
-                })
-            }
-            else {
-                var uploadFile = new File();
-
-                uploadFile.originalname = req.file.originalname;
-                uploadFile.encoding = req.file.encoding;
-                uploadFile.mimetype = req.file.mimetype;
-                uploadFile.destination = req.file.destination;
-                uploadFile.filename = req.file.filename;
-                uploadFile.size = req.file.size;
-                uploadFile.user_id = req.user_id;
-
-                uploadFile.save(function (err, file) {
-                    if (err) {
-                        return res.status(500).send({
-                            success: false,
-                            message: 'Error: Server error'
-                        });
-                    }
-                    console.log(file._id + " Added to DB.");
-
-                    User.findOneAndUpdate({
-                        _id: req.user_id
-                    }, {
-                            $push: { files: file._id }
-                        }, { new: true }, function (err, user) {
-                            if (err) {
-                                return res.status(500).send({
-                                    success: false,
-                                    message: 'Error: Server error'
-                                });
-                            }
-                            else {
-                                console.log("File added to user " + user._id);
-                                return res.status(200).send({
-                                    success: true,
-                                    message: "File uploaded and added to DB",
-                                    data: file
-                                });
-                            }
-                        });
-                });
-            }
-        })
-
     });
 
-    app.get('/api/admin/file/:filename', requireRole('admin'), function (req, res) {
+    app.get('/api/admin/file/:filename', requireRole('admin'), retrieveFile(dir) ,function (req, res) {
         if (!req.params.filename) {
             return res.status(400).send({
                 success: false,
                 message: "Error: filename has not been entered in parameters"
             });
         }
-        File.find({
-            originalname: req.params.filename
-        }, function (err, files) {
-            if (err) {
-                return res.status(500).send({
-                    success: false,
-                    message: "Error: server error"
-                });
-            }
-            else if (files.length == 0) {
-                return res.status(404).send({
-                    success: false,
-                    message: "Error: No file found with this name"
-                });
-            }
-            var file = files[0];
-            var filePath = path.join(dir, file.originalname);
-            var stream = fs.createReadStream(filePath);
-            stream.on('error', function (error) {
-                res.writeHead(404, 'Not Found');
-                res.end();
-            });
-            stream.pipe(res);
-            //TODO: Make file downloadable
-            //TODO: Delete file endpoint
-        });
     });
 }
